New API integrated, Weapon/Armor calls mostly functional
You can now make API calls for Armor and Weapons which use the new API. Text needs to be added to functions to describe the attributes shown and the last two properties on the armor tab need to be fixed to actually show their values. Attempted to make api call work on button press in the background page but need more work on that.